% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_propulmeta.R
\docType{data}
\name{cisapride}
\alias{cisapride}
\title{Cisapride and Reflux Esophagitis Dataset}
\format{
A data frame with 13 rows and 5 variables:
\describe{
\item{study}{Study label}
\item{event.cisa}{Number of healed patients in the cisapride group}
\item{n.cisa}{Total number of patients in the cisapride group}
\item{event.plac}{Number of healed patients in the placebo group}
\item{n.plac}{Total number of patients in the placebo group}
}
}
\source{
Adapted from \code{meta} package example datasets.
}
\usage{
data(cisapride)
}
\description{
This dataset provides data from randomized controlled trials evaluating the efficacy
of Cisapride versus placebo in healing reflux esophagitis.
}
\details{
This dataset is used to illustrate meta-analytic methods for binary outcomes like risk ratios and odds ratios.
}
\examples{
data(cisapride)
head(cisapride)
}
\keyword{datasets}
