% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_propulmeta.R
\docType{data}
\name{lungcancer}
\alias{lungcancer}
\title{Lung Cancer and Smoking Cohort Study Dataset}
\format{
A data frame with 7 rows and 6 variables:
\describe{
\item{study}{Study label}
\item{participants}{Number of participants in the study}
\item{d.smokers}{Number of lung cancer deaths among smokers}
\item{py.smokers}{Person-years among smokers}
\item{d.nonsmokers}{Number of lung cancer deaths among non-smokers}
\item{py.nonsmokers}{Person-years among non-smokers}
}
}
\source{
Adapted from \code{meta} package examples.
}
\usage{
data(lungcancer)
}
\description{
Data from cohort studies investigating the association between smoking and lung cancer.
}
\details{
This dataset is often used for meta-analysis of cohort study data to calculate relative risks (RR) for smoking.
}
\examples{
data(lungcancer)
head(lungcancer)
}
\keyword{datasets}
